requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiofiles==23.2.1
alabaster==0.7.16
albumentations==1.4.0
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.7.0
boto3==1.34.86
botocore==1.34.86
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
compel==2.0.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
deepface==0.0.91
dependency-injector==4.41.0
diffusers==0.26.3
docconvert==2.1.0
docker-pycreds==0.4.0
docutils==0.20.1
easydict==1.12
easygui==0.98.3
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
gdown==5.1.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.27.0
gradio_client==0.15.1
griffe==0.41.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib_resources==6.4.0
insightface==0.7.3
ipython==8.22.2
ipywidgets==8.1.2
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lit==15.0.7
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==1.0.1
mkdocs-material==9.5.13
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.1
mkdocstrings-python==1.8.0
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
packaging==23.2
paginate==0.5.6
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
peft==0.11.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pyfacer==0.0.4
Pygments==2.17.2
pymdown-extensions==10.7.1
pyparsing==3.1.1
pyre-extensions==0.0.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
qudida==0.0.4
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.1
s3transfer==0.10.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorrt==8.6.1.post1
termcolor==2.4.0
tf_keras==2.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.16
tinydb==4.8.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.2
traitlets==5.14.2
transformers==4.25.1
triton==2.0.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.15
urllib3==2.2.1
uvicorn==0.29.0
validators==0.22.0
voluptuous==0.14.2
wandb==0.16.4
watchdog==4.0.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
xformers==0.0.20
zipp==3.17.0
onnxruntime-gpu==1.16.1
lpips== 0.1.4
